fix: do not update on getUser
