+++ src/api/analyze/controllers/getGraph.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { existsSync } from 'fs';
 import { logger } from '~/config/winston';
-import { generateAeryGraph } from '~/utils/generateAeryGraph';
+import { generateAeryGraph } from '~/utils/aery/generateAeryGraph';
 
 export const getGraph = async (req: Request, res: Response) => {
     try {
