+++ src/api/analyze/index.ts
@@ -2,6 +2,6 @@ import express, { Router } from 'express';
 
 const router: Router = express.Router();
 
-router.use('/analyze', require('~/api/analyze/routes/graph'));
+router.use('/analyze', require('~/api/analyze/routes/analyze'));
 
 module.exports = router;
